+++ pysys-examples/internal/testcases/PySys_internal_001/run.py
@@ -12,7 +12,7 @@ class PySysTest(BaseTest):
 
 	def execute(self):
 		
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 		try:
 			runPySys(self, 'pysys', ['run', '-o', 'myoutdir', 
 				'-X', 'projectbooloverride=fAlse',