+++ pysys-examples/internal/testcases/PySys_internal_067/run.py
@@ -10,7 +10,7 @@ from pysysinternalhelpers import *
 class PySysTest(BaseTest):
 
 	def execute(self):
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 
 		runPySys(self, 'pysys', ['run', '-o', self.output+'/myoutdir', '-v', 'DEBUG'], workingDir='test', ignoreExitStatus=True, 
 			environs={