+++ pysys-examples/internal/testcases/PySys_internal_068/run.py
@@ -5,8 +5,8 @@ from pysys.constants import *
 from pysys.basetest import BaseTest
 import os, sys, math, shutil, glob, locale
 
-# contains a non-ascii £ character that is different in utf-8 vs latin-1
-TEST_STR = u'Hello £ world' 
+# contains a non-ascii ï¿½ character that is different in utf-8 vs latin-1
+TEST_STR = u'Hello ï¿½ world' 
 
 if PROJECT.testRootDir+'/internal/utilities/extensions' not in sys.path:
 	sys.path.append(PROJECT.testRootDir+'/internal/utilities/extensions') # only do this in internal testcases; normally sys.path should not be changed from within a PySys test
@@ -18,7 +18,7 @@ class PySysTest(BaseTest):
 		
 		if locale.getpreferredencoding() in ['ANSI_X3.4-1968', 'ascii']: self.skipTest('cannot run in ASCII locale')
 
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 		# make testRootDir and working dir be different
 		os.rename(self.output+'/test/pysysproject.xml', self.output+'/pysysproject.xml')