+++ pysys-examples/internal/testcases/PySys_internal_076/run.py
@@ -11,7 +11,7 @@ from pysysinternalhelpers import *
 class PySysTest(BaseTest):
 
 	def execute(self):
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 		runPySys(self, 'pysys', ['run', '-o', self.output+'/output', '--purge', '-v', 'debug'], workingDir='test')