+++ pysys-examples/internal/testcases/PySys_internal_078/run.py
@@ -11,7 +11,7 @@ from pysysinternalhelpers import *
 class PySysTest(BaseTest):
 
 	def execute(self):
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 		runPySys(self, 'all', ['print'], workingDir='test')
 		runPySys(self, 'full', ['print', '--full'], workingDir='test')
 		runPySys(self, 'include', ['print', '--include', 'group1', '--include', 'group2'], workingDir='test')