+++ pysys-examples/internal/testcases/PySys_internal_085/run.py
@@ -6,8 +6,8 @@ from pysys.basetest import BaseTest
 import os, sys, math, shutil, glob
 import locale
 
-# contains a non-ascii £ character that is different in utf-8 vs latin-1
-TEST_STR = u'Hello £ world' 
+# contains a non-ascii ï¿½ character that is different in utf-8 vs latin-1
+TEST_STR = u'Hello ï¿½ world' 
 
 if PROJECT.testRootDir+'/internal/utilities/extensions' not in sys.path:
 	sys.path.append(PROJECT.testRootDir+'/internal/utilities/extensions') # only do this in internal testcases; normally sys.path should not be changed from within a PySys test
@@ -19,7 +19,7 @@ class PySysTest(BaseTest):
 		
 		if locale.getpreferredencoding() in ['ANSI_X3.4-1968', 'ascii']: self.skipTest('cannot run in ASCII locale')
 
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 
 		runPySys(self, 'pysys', ['run', '-o', self.output+'/myoutdir', '--record', '--cycle', '2', '-n', '2'], ignoreExitStatus=True, workingDir='test')
 		self.logFileContents('pysys.out', maxLines=0)