+++ pysys-examples/internal/testcases/PySys_internal_112/run.py
@@ -11,7 +11,7 @@ from pysysinternalhelpers import *
 class PySysTest(BaseTest):
 
 	def execute(self):
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 
 		runPySys(self, 'out-all', ['run', '--mode=ALL', '--record'], workingDir='test')