+++ pysys-examples/internal/testcases/PySys_internal_115/run.py
@@ -10,7 +10,7 @@ from pysysinternalhelpers import *
 
 class PySysTest(BaseTest):
 	def execute(self):
-		shutil.copytree(self.input, self.output+'/test')
+		self.copy(self.input, self.output+'/test')
 
 		runPySys(self, 'run-relative-outdir', 
 			['run', '--outdir', 'outdir-relative', '--mode=ALL'], workingDir='test')